romario/requirements.txt
django
django-stubs
debugpy
djangorestframework
django-browser-reload
django-environ
django-tailwind-4
psycopg[binary,pool]
pytest
pytest-cov
pytest-django
pytest-mock
pylint
pyright
bandit
junit2html
pysonar-scanner
pandas
reportlab
odfpy
django-phonenumber-field
django-import-export
tablib
openpyxl
unidecode
djangorestframework-simplejwt
django-cors-headers
httpie

